Refactor loading skeletons for module pages to use SSR and refactor components for maintainability
