+++ app/layout.tsx
@@ -1,21 +1,27 @@
 import React from "react";
-import { ClerkProvider } from "@clerk/nextjs";
-import { dark } from '@clerk/themes'
-import "./globals.css";
 import type { Metadata, Viewport } from "next";
-import ClientSidebar from "@/components/ClientSidebar";
-import { Providers as AppProviders } from "@/lib/providers";
-import { SidebarProvider } from "@/lib/sidebar-context";
-import { SessionProvider } from "@/context/SessionContext";
-import AnonymousDataMigration from "@/components/AnonymousDataMigration";
-import Header from "@/components/Header";
-import localFont from 'next/font/local'
+
+import { dark } from "@clerk/themes";
+import "./globals.css";
+import localFont from "next/font/local";
+
+import ClientSidebar from "@/components/Main/ClientSidebar";
+
+import { ClerkProvider } from "@clerk/nextjs";
+import { Providers as AppProviders } from "@/context/providers";
+import { SidebarProvider } from "@/context/sidebar-context";
+import { SessionProvider } from "@/context/session-context";
+
+import AnonymousDataMigration from "@/components/dialogs/AnonymousDataMigration";
+import Header from "@/components/Main/Header";
 
 // Session initializer - client component
-import { SessionInitializer } from "@/components/SessionInitializer";
+import { SessionInitializer } from "@/components/Main/SessionInitializer";
 
 // Font files
-const myFont = localFont({ src: 'fonts/AtkinsonHyperlegibleNextVF-Variable.woff2' })
+const myFont = localFont({
+  src: "fonts/AtkinsonHyperlegibleNextVF-Variable.woff2",
+});
 
 export const metadata: Metadata = {
   title: "Study Chat - Your Personal Study Assistant",
@@ -28,7 +34,6 @@ export const viewport: Viewport = {
   themeColor: "#ffffff",
 };
 
-
 // Clerk appearance config
 const clerkAppearance = {
   baseTheme: [dark],
@@ -38,13 +43,12 @@ const clerkAppearance = {
   },
   elements: {
     modalCloseButton: "absolute right-4 top-4",
-    card: "rounded-lg shadow-md",    
+    card: "rounded-lg shadow-md",
     formButtonPrimary:
       "bg-blue-600 hover:bg-blue-700 text-white font-semibold py-2 px-4 rounded",
     footerActionLink: "text-blue-600 hover:text-blue-800 font-semibold",
   },
 };
-
 
 export default function RootLayout({
   children,
