+++ app/modules/page.tsx
@@ -1,492 +1,25 @@
-"use client";
-
-import React, { Suspense, useState } from "react";
-import { useAuth } from "@clerk/nextjs";
-import Link from "next/link";
-import { Plus, Search } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Input } from "@/components/ui/input";
-import { SignInButton } from "@clerk/nextjs";
-import { formatDate } from "@/lib/utils";
-import { useEffect } from "react";
-import { encodeModuleSlug } from "@/lib/utils";
-import { ResourceUploadButton } from "@/components/ResourceUploadButton";
-import { useSession } from "@/context/SessionContext";
-import { api } from "@/lib/api";
-import { ResourceTable } from "@/components/ResourceTable";
-
-// Import the dedicated search params component
-import { SearchParamsReader } from "./search-params";
-
-// Client component for module operations to be loaded in a Suspense boundary
-import ModuleOperations from "./module-operations";
-
-// Define interfaces for the module data
-interface Module {
-  id: string;
-  name: string;
-  description: string | null;
-  icon: string;
-  resourceCount: number;
-  updatedAt: string;
-}
-
-// Define interface for the resource data
-interface Resource {
-  id: string;
-  title: string;
-  description: string;
-  type: string;
-  url?: string | null;
-  moduleId: string;
-  moduleName?: string | null;
-  createdAt: string;
-  _deleted?: boolean;
-}
+import ModulesPageWrapper from "@/components/Modules/modules-page-wrapper";
 
 /**
- * Renders a loading skeleton for the modules page.
+ * Renders the modules page with search parameters handling.
  *
- * This component displays a placeholder UI that mimics the layout of the modules page while data is loading.
- * It includes a header for "Categories", static tabs for "Modules" and "All Resources", a search bar skeleton,
- * and a disabled "Create Module" button with a Plus icon.
+ * This server component forces a short delay to ensure the loading state is properly displayed.
+ * It then renders the client component which handles the actual data fetching and UI.
  *
- * @returns The React element representing the loading skeleton layout.
+ * @returns A React element representing the complete modules page with dynamic content.
  */
-function ModulesLoading() {
+export default async function ModulesPage() {
+  // This artificial delay ensures that the loading.tsx is displayed
+  // when navigating to this page
+  await new Promise((resolve) => setTimeout(resolve, 1000));
+
   return (
     <div className="flex min-h-screen w-full flex-col">
-      <div className="flex-1 space-y-4">
-        <div className="flex items-center justify-between px-3 py-3.5">
-          <h1 className="font-bold text-xl">Categories</h1>
-        </div>
-
-        {/* Tabs with static labels */}
-        <div className="px-3">
-          <div className="grid w-full max-w-md grid-cols-2 h-9 bg-gray-100 rounded-lg">
-            <div className="flex items-center justify-center m-1 h-7 bg-white rounded-lg">
-              <span className="font-medium text-sm">Modules</span>
-            </div>
-            <div className="flex items-center justify-center m-1 h-7 w-auto rounded-sm">
-              <span className="font-medium text-sm text-muted-foreground">
-                All Resources
-              </span>
-            </div>
-          </div>
-
-          {/* Search bar and button skeleton */}
-          <div className="relative flex items-center gap-2 mt-4 mb-6">
-            <div className="relative flex-1 max-w-md">
-              <div className="h-9 w-full bg-gray-200 animate-pulse rounded"></div>
-            </div>
-            <Button variant="outline" disabled>
-              <Plus className="h-4 w-4 mr-2" />
-              Create Module
-            </Button>
-          </div>
-
-          {/* Empty area for modules to load into */}
-          <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3 mt-2">
-            {/* No skeleton cards as requested */}
-          </div>
-        </div>
-      </div>
+      <ModulesPageWrapper />
     </div>
   );
 }
 
-/**
- * Renders the modules page content with tabs for modules and resources.
- *
- * This component fetches and displays modules based on an optional search query, using URL parameters to
- * determine the active tab, whether to open the resource upload dialog, and a preselected module for uploading resources.
- * It checks for user authentication, displaying a sign-in prompt if the user is unauthorized, and conditionally renders
- * either a filtered list of modules or a resource table through a nested component.
- *
- * @param searchParams - URL query parameters that set the initial UI state (e.g., active tab, resource upload dialog visibility, and preselected module).
- */
-function ModulesPageContent({
-  searchParams,
-}: {
-  searchParams: URLSearchParams;
-}) {
-  const { isLoaded, isSignedIn } = useAuth();
-  const { sessionId, isLoading: sessionLoading } = useSession();
-  const [searchQuery, setSearchQuery] = useState("");
-  const [modules, setModules] = useState<Module[]>([]);
-  const [filteredModules, setFilteredModules] = useState<Module[]>([]);
-  const [activeTab, setActiveTab] = useState(() => {
-    // Get tab from URL param or default to "modules"
-    return searchParams?.get("tab") || "modules";
-  });
-  const [isLoading, setIsLoading] = useState(true);
-  const shouldOpenResourceUpload =
-    searchParams?.get("openResourceUpload") === "true";
-  const preselectedModuleId = searchParams?.get("moduleId");
-  const [openResourceUpload, setOpenResourceUpload] = useState(
-    shouldOpenResourceUpload
-  );
-
-  // This effect fetches the modules data from the API
-  useEffect(() => {
-    async function fetchModules() {
-      try {
-        setIsLoading(true);
-
-        const data = await api.getModules(searchQuery || undefined);
-
-        // Extract modules from the response object
-        const modulesList = data.modules || [];
-        setModules(modulesList);
-        setFilteredModules(modulesList);
-      } catch (error) {
-        console.error("Error fetching modules:", error);
-        // Set modules to empty array on error
-        setModules([]);
-        setFilteredModules([]);
-      } finally {
-        setIsLoading(false);
-      }
-    }
-
-    if (isLoaded && !sessionLoading) {
-      fetchModules();
-    }
-  }, [isSignedIn, isLoaded, sessionId, sessionLoading, searchQuery]);
-
-  // Filter modules based on search query whenever searchQuery changes
-  useEffect(() => {
-    if (activeTab === "modules" && modules.length > 0) {
-      if (!searchQuery) {
-        setFilteredModules(modules);
-      } else {
-        const filtered = modules.filter(
-          (module) =>
-            module.name.toLowerCase().includes(searchQuery.toLowerCase()) ||
-            (module.description &&
-              module.description
-                .toLowerCase()
-                .includes(searchQuery.toLowerCase()))
-        );
-        setFilteredModules(filtered);
-      }
-    }
-  }, [searchQuery, modules, activeTab]);
-
-  // Add to existing useEffect that loads modules
-  useEffect(() => {
-    // Handle opening the resource upload dialog from URL params
-    if (shouldOpenResourceUpload) {
-      setOpenResourceUpload(true);
-      // Remove the query parameter from the URL without refreshing the page
-      const newUrl = new URL(window.location.href);
-      newUrl.searchParams.delete("openResourceUpload");
-      window.history.replaceState({}, "", newUrl.toString());
-    }
-  }, [shouldOpenResourceUpload]);
-
-  if (!isLoaded || sessionLoading) {
-    return <ModulesLoading />;
-  }
-
-  // Check if the user can access modules (signed in or has sessionId)
-  const canAccessModules = isSignedIn || !!sessionId;
-
-  if (!canAccessModules) {
-    return (
-      <div className="flex min-h-screen w-full flex-col">
-        <div className="flex-1 space-y-4">
-          <div className="flex p-5 items-center">
-            <h1 className="font-bold text-xl">Categories</h1>
-          </div>
-          <div className="flex flex-col items-center justify-center space-y-4 rounded-lg border border-dashed p-8 text-center">
-            <h3 className="text-xl font-medium">Please sign in</h3>
-            <p className="text-muted-foreground">
-              You need to be signed in to view and manage your modules
-            </p>
-            <SignInButton mode="modal">
-              <Button>Sign in</Button>
-            </SignInButton>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  /**
-   * Fetches modules and resources from the API and displays them in a resource table, filtered by a search query.
-   *
-   * This component retrieves module data for selector options and, if the user is signed in, fetches resource data.
-   * It filters the resources by matching the search query against their title, description, or associated module name.
-   * Additionally, it handles resource updates by marking deleted resources appropriately.
-   *
-   * @param searchQuery - A query string used to filter the displayed resources.
-   *
-   * @returns A JSX element rendering the resource table with the filtered resources.
-   */
-  function ResourcesWrapper({ searchQuery }: { searchQuery: string }) {
-    const [filteredResources, setFilteredResources] = useState<Resource[]>([]);
-    const [modules, setModules] = useState<
-      { id: string; name: string; icon: string }[]
-    >([]);
-    const [resourcesLoading, setResourcesLoading] = useState(true);
-    const { sessionId } = useSession();
-    const { isSignedIn } = useAuth();
-
-    // Fetch resources from the API
-    useEffect(() => {
-      /**
-       * Fetches modules and, if the user is authenticated, resources data from the API, updating state accordingly.
-       *
-       * This asynchronous function begins by setting the resources loading state to true and retrieves all modules, which are then simplified
-       * to include only the id, name, and icon. If the user is signed in, it attempts to fetch resources from the "/api/resources" endpoint.
-       * For unauthorized responses (HTTP 401), it logs a message and clears the filtered resources. If the resources fetch fails for other reasons,
-       * an error is thrown, caught, and logged, with the filtered resources reset to an empty array. Finally, the loading state is set to false.
-       *
-       * @remark
-       * Resources are filtered by the search query if provided; otherwise, all fetched resources are used.
-       */
-      async function fetchData() {
-        try {
-          setResourcesLoading(true);
-
-          // Fetch all modules for the selector
-          const modulesData = await api.getModules();
-          const modulesList = modulesData.modules || [];
-          setModules(
-            modulesList.map((m: Module) => ({
-              id: m.id,
-              name: m.name,
-              icon: m.icon,
-            }))
-          );
-
-          // Only attempt to fetch resources if user is signed in
-          if (isSignedIn) {
-            // Fetch resources - these require authentication
-            const resourcesResponse = await fetch("/api/resources");
-
-            if (resourcesResponse.status === 401) {
-              // Handle unauthorized gracefully - user is not logged in
-              console.log("User is not authenticated for resources");
-              setFilteredResources([]);
-            } else if (!resourcesResponse.ok) {
-              throw new Error(
-                `Failed to fetch resources: ${resourcesResponse.statusText}`
-              );
-            } else {
-              const resourcesData = await resourcesResponse.json();
-
-              // Initialize filtered resources
-              if (!searchQuery) {
-                setFilteredResources(resourcesData);
-              } else {
-                const filtered = resourcesData.filter(
-                  (resource: Resource) =>
-                    resource.title
-                      .toLowerCase()
-                      .includes(searchQuery.toLowerCase()) ||
-                    (resource.description &&
-                      resource.description
-                        .toLowerCase()
-                        .includes(searchQuery.toLowerCase())) ||
-                    (resource.moduleName &&
-                      resource.moduleName
-                        .toLowerCase()
-                        .includes(searchQuery.toLowerCase()))
-                );
-                setFilteredResources(filtered);
-              }
-            }
-          } else {
-            // User is not signed in, don't try to fetch resources
-            setFilteredResources([]);
-          }
-        } catch (error) {
-          console.error("Error fetching data:", error);
-          // Set resources to empty array on error
-          setFilteredResources([]);
-        } finally {
-          setResourcesLoading(false);
-        }
-      }
-
-      fetchData();
-    }, [searchQuery, isSignedIn, sessionId]);
-
-    const handleResourceUpdate = (updatedResource: Resource) => {
-      if (updatedResource._deleted) {
-        // If resource was deleted, mark as deleted in the UI
-        setFilteredResources((resources) =>
-          resources.map((r) =>
-            r.id === updatedResource.id ? { ...r, _deleted: true } : r
-          )
-        );
-      } else {
-        // Regular update
-        setFilteredResources((resources) =>
-          resources.map((r) =>
-            r.id === updatedResource.id ? updatedResource : r
-          )
-        );
-      }
-    };
-
-    return (
-      <div className="mt-6">
-        <div className="flex items-center justify-between mb-4">
-          {searchQuery && (
-            <p className="text-muted-foreground">
-              Showing results for: &quot;{searchQuery}&quot;
-            </p>
-          )}
-        </div>
-
-        {/* Use the new ResourceTable component */}
-        <ResourceTable
-          resources={filteredResources}
-          modules={modules}
-          onUpdate={handleResourceUpdate}
-          showModuleColumn={true}
-          isLoading={resourcesLoading}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex min-h-screen w-full flex-col">
-      <div className="flex-1 space-y-4">
-        <div className="flex items-center justify-between px-3 py-3.5">
-          <h1 className="font-bold text-xl">Categories</h1>
-        </div>
-
-        {/* Tabs for Module and Resources - now at the top */}
-        <Tabs
-          defaultValue={activeTab}
-          className="px-3"
-          onValueChange={setActiveTab}
-        >
-          <TabsList className="grid w-full max-w-md grid-cols-2">
-            <TabsTrigger value="modules">Modules</TabsTrigger>
-            <TabsTrigger value="resources">All Resources</TabsTrigger>
-          </TabsList>
-
-          {/* Search bar with contextual button - now below tabs */}
-          <div className="relative flex items-center gap-2 mt-4 mb-6">
-            <div className="relative flex-1 max-w-md">
-              <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-4 w-4" />
-              <Input
-                placeholder={
-                  activeTab === "modules"
-                    ? "Search modules..."
-                    : "Search resources..."
-                }
-                className="pl-10"
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-              />
-            </div>
-            {/* Contextual button that changes based on active tab */}
-            {activeTab === "modules" ? (
-              <Suspense fallback={<Button disabled>Loading...</Button>}>
-                <ModuleOperations sessionId={sessionId} />
-              </Suspense>
-            ) : (
-              <ResourceUploadButton
-                variant="outline"
-                moduleId={preselectedModuleId || undefined}
-                initialOpen={openResourceUpload}
-              ></ResourceUploadButton>
-            )}
-          </div>
-
-          <TabsContent value="modules" className="mt-2">
-            {isLoading ? (
-              <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-                {/* No skeleton cards as requested */}
-              </div>
-            ) : filteredModules.length === 0 ? (
-              <div className="flex flex-col items-center justify-center space-y-4 rounded-lg border border-dashed p-8 text-center">
-                <h3 className="text-xl font-medium">No modules found</h3>
-                <p className="text-muted-foreground">
-                  {searchQuery
-                    ? "Try a different search term"
-                    : "Create your first module to get started"}
-                </p>
-                {!searchQuery}
-              </div>
-            ) : (
-              <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-                {filteredModules.map((module) => (
-                  <Link
-                    href={`/modules/${encodeModuleSlug(module.name)}`}
-                    key={module.name}
-                  >
-                    <Card className="hover:bg-muted/50 transition-colors">
-                      <CardHeader>
-                        <CardTitle>
-                          <span className="mr-2">{module.icon}</span>
-                          {module.name}
-                        </CardTitle>
-                        <CardDescription>
-                          {module.description || "No description"}
-                        </CardDescription>
-                      </CardHeader>
-                      <CardContent>
-                        <div className="text-sm text-muted-foreground">
-                          {module.resourceCount}{" "}
-                          {module.resourceCount === 1
-                            ? "resource"
-                            : "resources"}
-                        </div>
-                      </CardContent>
-                      <CardFooter className="text-xs text-muted-foreground">
-                        Updated {formatDate(module.updatedAt)}
-                      </CardFooter>
-                    </Card>
-                  </Link>
-                ))}
-              </div>
-            )}
-          </TabsContent>
-
-          <TabsContent value="resources" className="mt-2">
-            <ResourcesWrapper searchQuery={searchQuery} />
-          </TabsContent>
-        </Tabs>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Renders the main modules page using nested Suspense boundaries.
- *
- * The component wraps the ModulesPageContent inside two Suspense boundaries, each using ModulesLoading as a fallback.
- * It utilizes a SearchParamsReader to extract URL search parameters and passes them to ModulesPageContent to manage
- * module display and resource loading.
- */
-export default function ModulesPage() {
-  return (
-    <Suspense fallback={<ModulesLoading />}>
-      <Suspense fallback={<ModulesLoading />}>
-        <SearchParamsReader>
-          {(searchParams) => <ModulesPageContent searchParams={searchParams} />}
-        </SearchParamsReader>
-      </Suspense>
-    </Suspense>
-  );
-}
-
-// Add this export to allow dynamic rendering
+// This tells Next.js to always render the page dynamically
+// which helps trigger the loading state properly
 export const dynamic = "force-dynamic";
