+++ context/providers.tsx
@@ -3,7 +3,7 @@
 import { ThemeProvider as NextThemesProvider } from "next-themes";
 import type { ThemeProviderProps } from "next-themes";
 import { Toaster } from "sonner";
-import { SessionProvider } from "@/context/SessionContext";
+import { SessionProvider } from "@/context/session-context";
 
 export function Providers({ children, ...props }: ThemeProviderProps) {
   return (
